Move appmanifest related stuff to separate file
