+++ src/utils/helpers.ts
@@ -86,20 +86,3 @@ export const decode = (string: string) =>
     Buffer.from(string, CONTENT_ENCODINGS.base64 as BufferEncoding).toString(CONTENT_ENCODINGS.utf8 as BufferEncoding);
 export const encode = (string: string) =>
     Buffer.from(string, CONTENT_ENCODINGS.utf8 as BufferEncoding).toString(CONTENT_ENCODINGS.base64 as BufferEncoding);
-
-const isLegacyAppManifest = (decodedAppManifestContent: any) => decodedAppManifestContent instanceof Array;
-
-export const updateAppManifestContent = (decodedAppManifestContent: any, appName: string, vspecPath: string, dataPoints: any[]) => {
-    if (isLegacyAppManifest(decodedAppManifestContent)) {
-        // for backwards compatibility to AppManifest v2
-        decodedAppManifestContent[0].name = appName.toLowerCase();
-        decodedAppManifestContent[0].vehicleModel.src = vspecPath;
-        decodedAppManifestContent[0].vehicleModel.datapoints = dataPoints;
-    } else {
-        const vsiIndex = decodedAppManifestContent.interfaces.findIndex((entry: any) => entry.type === 'vehicle-signal-interface');
-        decodedAppManifestContent.name = appName.toLowerCase();
-        decodedAppManifestContent.interfaces[vsiIndex].config.src = vspecPath;
-        decodedAppManifestContent.interfaces[vsiIndex].config.datapoints.required = dataPoints;
-    }
-    return decodedAppManifestContent;
-};
